Project/Mobile/Skin_Check_App_v2/Project_1/skin_check_api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
numpy>=1.19.2
tensorflow==2.16.1
h5py>=3.1.0
